+++ src/components/FamilyFriday.js
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from "react";
 import Button from "./Button";
 import LunchGroups from "./LunchGroups";
-import NewHire from "./NewHire";
-import { getEmployees, resetEmployees } from "../functions/api";
+import Roster from "./Roster";
+import { getEmployees } from "../functions/api";
 
 export default function FamilyFriday() {
   const [employees, setEmployees] = useState();
@@ -16,30 +16,10 @@ export default function FamilyFriday() {
     setGroups();
   }, [employees]);
 
-  const reset = () => {
-    resetEmployees().then(res => setEmployees(res));
-  };
-
   return (
     <div className="familyFriday">
       <h1>Family Friday Lunch Lottery</h1>
-      {employees ? (
-        <React.Fragment>
-          <NewHire
-            addEmployee={employee => setEmployees([...employees, employee])}
-          />
-          <Button
-            onClick={reset}
-            label={"Reset Employees"}
-            title={"Reset Employees"}
-          />
-        </React.Fragment>
-      ) : (
-        "Loading Employees..."
-      )}
-      <div>
-        {employees && employees.map(employee => employee.name).join(", ")}
-      </div>
+      <Roster employees={employees} setEmployees={setEmployees} />
 
       <Button
         onClick={() => setGroups(getGroups(employees))}
